% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/082_connect_sgbd.R
\name{connect_sgbd}
\alias{connect_sgbd}
\title{Função que cria a conexão com o Banco de Dados}
\usage{
connect_sgbd(armazenar)
}
\arguments{
\item{armazenar}{Indica a forma de armazenamento dos dados: 'csv' ou 'sqlite'}
}
\description{
Função que cria a conexão com o Banco de Dados
}
