% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_gerar_bd_cnpj.R
\name{gerar_bd_cnpj}
\alias{gerar_bd_cnpj}
\title{Função que orquestra as demais funções para realizar o tratamento e organização dos dados do CNPJ}
\usage{
gerar_bd_cnpj(
  path_arquivos_txt,
  localizar_cnpj = "NAO",
  n_lines = 1e+05,
  armazenar = "csv"
)
}
\arguments{
\item{path_arquivos_txt}{Caminho (path) dos arquivos com a base de dados do CNPJ no formato '.txt'}

\item{localizar_cnpj}{Vetor com o número dos CNPJ que se deseja filtrar e obter os dados.
O valor padrão é "NAO", o que força ao tratamento de todas as linha da base de dados}

\item{n_lines}{Número de linhas que podem ser iteradas por vez: 10000, 100000 ou 1000000}

\item{armazenar}{Indica a forma de armazenamento dos dados: 'csv' ou 'sqlite' (OBS1: O delimitador do CSV é o simbolo: "#'),
(OBS2: Preferencialmente, defina a pasta de trabalho da sessão 'Working Directory' na mesma em que está localizado o arquivo
'.txt' da base de dados no CNPJ)}
}
\description{
Essa função foi desenvolvida utilizando como elementro central uma função complementar
chamada 'readr::read_lines_chunked', com o propósito de ler o arquivo '.txt' de 87Gb em partes de
10.000, 100.000 ou 1.000.000 de linhas por vez.
}
\examples{
\dontrun{
qsacnpj::gerar_bd_cnpj(path_arquivos_txt = "D:/qsa_cnpj",
                       localizar_cnpj = "NAO",
                       n_lines = 100000,
                       armazenar = "csv")


# Exemplo com número de CNPJ, entre aspas (""), do Banco do Brasil, Banco do Nordeste,
# Banco da Amazônia e Caixa Econômica

qsacnpj::gerar_bd_cnpj(path_arquivos_txt = "D:/qsa_cnpj",
                       localizar_cnpj = c("00000000000191", "07237373000120",
                                            "00360305000104", "04902979000144"),
                       n_lines = 100000,
                       armazenar = "sqlite")
}


}
