% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_obter_dados_cnae.R
\name{obter_dados_cnae}
\alias{obter_dados_cnae}
\title{Função que inclui a Tabela com os CNAEs}
\usage{
obter_dados_cnae(armazenar)
}
\arguments{
\item{armazenar}{Indica a forma de armazenamento dos dados: 'csv' ou 'sqlite' (OBS1: O delimitador do CSV é o simbolo: "#')}
}
\description{
Função que inclui a Tabela com os CNAEs
}
