% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_obter_dados_codigo_municipio_siafi.R
\name{obter_dados_codigo_municipios_siafi}
\alias{obter_dados_codigo_municipios_siafi}
\title{Função que inclui a Tabela de Código dos Municípios do SIAFI, correspondentes ao Código dos Municípios do IBGE}
\usage{
obter_dados_codigo_municipios_siafi(armazenar)
}
\arguments{
\item{armazenar}{Indica a forma de armazenamento dos dados: 'csv' ou 'sqlite' (OBS1: O delimitador do CSV é o simbolo: "#')}
}
\description{
Função que inclui a Tabela de Código dos Municípios do SIAFI, correspondentes ao Código dos Municípios do IBGE
}
