% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_obter_dados_natureza_juridica.R
\name{obter_dados_natureza_juridica}
\alias{obter_dados_natureza_juridica}
\title{Função que a Tabela de Classificação da Natureza Jurídica}
\usage{
obter_dados_natureza_juridica(armazenar)
}
\arguments{
\item{armazenar}{Indica a forma de armazenamento dos dados: 'csv' ou 'sqlite' (OBS1: O delimitador do CSV é o simbolo: "#')}
}
\description{
Função que a Tabela de Classificação da Natureza Jurídica
}
