% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_obter_dados_qsa.R
\name{obter_dados_qsa}
\alias{obter_dados_qsa}
\title{Função que orquestra as demais funções para realizar o tratamento e organização dos dados}
\usage{
obter_dados_qsa(path_arquivos_txt, localizar_cnpj, n_lines, armazenar)
}
\arguments{
\item{path_arquivos_txt}{Caminho (path) dos arquivos com a base de dados do CNPJ.}

\item{localizar_cnpj}{Vetor com o número dos CNPJ que se deseja filtrar e obter os dados.
O valor padrão é "NAO", o que força ao tratamento de todas as linha da base de dados}

\item{n_lines}{Número de linhas que podem ser iteradas por vez: 10000, 100000 ou 1000000}

\item{armazenar}{Indica a forma de armazenamento dos dados: 'csv' ou 'sqlite' (OBS1: O delimitador do CSV é o simbolo: "#'),
(OBS2: Preferencialmente, defina a pasta de trabalho da sessão 'Working Directory' na mesma em que estão localizados os arquivos
da base de dados no CNPJ)}
}
\description{
Essa função foi desenvolvida utilizando como elementro central uma função complementar
chamada 'readr::read_lines_chunked', com o propósito de ler o arquivo de 95Gb em partes de
10.000, 100.000 ou 1.000.000 de linhas por vez.
}
