% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_obter_dados_qualificacao_responsavel.R
\name{obter_dados_qualificacao_responsavel}
\alias{obter_dados_qualificacao_responsavel}
\title{Função que a Tabela de Qualificação do Responsável no Quadro Societário da Pessoa Jurídica}
\usage{
obter_dados_qualificacao_responsavel(armazenar)
}
\arguments{
\item{armazenar}{Indica a forma de armazenamento dos dados: 'csv' ou 'sqlite' (OBS1: O delimitador do CSV é o simbolo: "#')}
}
\description{
Função que a Tabela de Qualificação do Responsável no Quadro Societário da Pessoa Jurídica
}
