% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_obter_dados_situacao_cadastral.R
\name{obter_dados_situacao_cadastral}
\alias{obter_dados_situacao_cadastral}
\title{Função que inclui a Tabela de Motivo Situação Cadastral da Pessoa Jurídica}
\usage{
obter_dados_situacao_cadastral(armazenar)
}
\arguments{
\item{armazenar}{Indica a forma de armazenamento dos dados: 'csv' ou 'sqlite' (OBS1: O delimitador do CSV é o simbolo: "#')}
}
\description{
Função que inclui a Tabela de Motivo Situação Cadastral da Pessoa Jurídica
}
