% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_pesquisar_dados_cnpj.R
\name{pesquisar_dados_cnpj}
\alias{pesquisar_dados_cnpj}
\title{Função que pesquisa os dados do CNPJ do Banco de Dados}
\usage{
pesquisar_dados_cnpj(cnpj_alvos, criar_indice = TRUE, sgbd = "sqlite")
}
\arguments{
\item{cnpj_alvos}{Números dos CNPJ que serão extraídos do Banco de Dados}

\item{criar_indice}{Cria um índice no Banco de Dados para otimizar as pesquisas.
OBS: Esse processo demora uns 5 min}

\item{sgbd}{Variável que indica o SGBD que será realizada a conexão. O 'sqlite' é a opção padrão.}
}
\description{
Ao final da consulta, é gerado um arquivo CSV com o resultado da pesquisa
}
