% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsacnpj-package.R
\docType{package}
\name{qsacnpj-package}
\alias{qsacnpj}
\alias{qsacnpj-package}
\title{qsacnpj: Pacote para tratar e organizar os dados do Cadastro Nacional da Pessoa Jurídica (CNPJ)}
\description{
O Cadastro Nacional da Pessoa Jurídica (CNPJ) é uma das bases de dados governamentais
        mais importantes do Brasil. Contudo, a Receita Federal do Brasil, infelizmente,
        disponibiliza a base de dados num formato que impede a sua utilização de imediato.
        Esse pacote ajuda no tratamento e organização dos dados.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/georgevbsantiago/qsacnpj}
}

}
\author{
\strong{Maintainer}: George Santiago \email{georgevbsantiago@hotmail.com}

}
\keyword{internal}
