% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_data.R
\docType{data}
\name{tab_cnpj_entes_publicos_br}
\alias{tab_cnpj_entes_publicos_br}
\title{Tabela com o CNPJ dos Entes Federativos}
\format{Um data frame com 118.287 linhas e 6 variáveis:
\describe{
\item{cnpj}{Código do CNPJ}
\item{cod_natureza_juridica}{Código da Natureza Jurídica}
\item{nome_empresarial}{Nome do Ente Federativo}
\item{ente_federativo_responsavel}{Código do Ente Federativo Responsável}
\item{ente_principal}{Indicador do Ente Principal}
\item{situacao}{Situação Cadastral}
...
}}
\source{
\url{http://www.receita.fazenda.gov.br/publico/cnpj/ListadosCNPJdosEntesFederativos.ods}
}
\usage{
tab_cnpj_entes_publicos_br
}
\description{
Dataset com o nome e código dos Entes Públicos Federais, Estaduais e Municipais
}
\keyword{datasets}
