% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_data.R
\docType{data}
\name{tab_codigo_municipios_siafi}
\alias{tab_codigo_municipios_siafi}
\title{Tabela de Código dos Municípios do SIAFI, correspondente ao do IBGE}
\format{Um data frame com 5589 linhas e 5 variáveis:
\describe{
\item{codigo_siafi}{Código do Município no SIAFI}
\item{cnpj}{CNPJ do Município}
\item{descricao}{Nome do Município}
\item{uf}{Unidade Federativa do Município}
\item{codigo_ibge}{Código do Município no IBGE}
...
}}
\source{
\url{http://www.tesourotransparente.gov.br/ckan/dataset/lista-de-municipios-do-siafi}
}
\usage{
tab_codigo_municipios_siafi
}
\description{
Dataset com o nome e código de Classificação dos Municípios no SIAFI, com correspondência ao do IGBE
}
\keyword{datasets}
