% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_data.R
\docType{data}
\name{tab_natureza_juridica}
\alias{tab_natureza_juridica}
\title{Tabela de Classificação da Natureza Jurídica}
\format{
Um data frame com 90 linhas e 4 variáveis:
\describe{
\item{cod_natureza_juridica}{Código da Natureza Jurídica}
\item{nm_natureza_juridica}{Nome da Natureza Jurídica}
\item{cod_subclass_natureza_juridica}{Código da subclassificação da Natureza Jurídica}
\item{nm_subclass_natureza_juridica}{Nome da subclassificação da Natureza Jurídica}
...
}
}
\source{
\url{https://concla.ibge.gov.br/estrutura/natjur-estrutura/natureza-juridica-2018}
}
\usage{
tab_natureza_juridica
}
\description{
Dataset com o nome e código de Classificação da Natureza Jurídica de 2018
}
\keyword{datasets}
