% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_data.R
\docType{data}
\name{tab_situacao_cadastral}
\alias{tab_situacao_cadastral}
\title{Tabela de Situação Cadastral}
\format{Um data frame com 68 linhas e 2 variáveis:
\describe{
\item{cod_situacao_cadastral}{Código da Situação Cadastral}
\item{nm_situacao_cadastral}{Nome da Situação Cadastral}
...
}}
\source{
\url{http://receita.economia.gov.br/orientacao/tributaria/cadastros/cadastro-nacional-de-pessoas-juridicas-cnpj/DominiosMotivoSituaoCadastral.csv}
}
\usage{
tab_situacao_cadastral
}
\description{
Dataset com o nome e código da Situação Cadastral da Pessoa Jurídica
}
\keyword{datasets}
