% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsa_cnpj.R
\name{tratar_arquivo_txt}
\alias{tratar_arquivo_txt}
\title{Função que cria as Tabelas do Banco de Dados}
\usage{
tratar_arquivo_txt(x, pos)
}
\arguments{
\item{x}{Data Frame que será analisado e tratado}

\item{pos}{Variável que indica o número da linha inicial da interação}
}
\description{
Função que cria as Tabelas do Banco de Dados
}
