% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_data.R
\docType{data}
\name{tab_qualificacao_responsavel_socio}
\alias{tab_qualificacao_responsavel_socio}
\title{Tabela de Qualificação do Sócio}
\format{Um data frame com 79 linhas e 3 variáveis:
\describe{
\item{cod_qualificacao_responsavel_socio}{Código da Qualificação do Responsável e/ou Sócio}
\item{nm_qualificacao_responsavel_socio}{Nome da Qualificação do Responsável e/ou Sócio}
\item{coletado_atualmente}{Informaçquão que indica se a Receita Federal utiliza ou coleta atualmente a classificação}
...
}}
\source{
\url{http://receita.economia.gov.br/orientacao/tributaria/cadastros/cadastro-nacional-de-pessoas-juridicas-cnpj/DominiosQualificaodoresponsvel.csv}
}
\usage{
tab_qualificacao_responsavel_socio
}
\description{
Dataset com o nome e código de Qualificação do Responsável e/ou Sócio no Quadro Societário da Pessoa Jurídica
}
\keyword{datasets}
