% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_obter_dados_qsa.R
\name{tratar_arquivo_txt}
\alias{tratar_arquivo_txt}
\title{Função trata e armazena os dados do CNPJ no Banco de Dados}
\usage{
tratar_arquivo_txt(path_arquivo_txt, localizar_cnpj, n_lines, armazenar)
}
\arguments{
\item{x}{Data Frame que será analisado e tratado}

\item{pos}{Variável que indica o número da linha inicial da interação}
}
\description{
Função trata e armazena os dados do CNPJ no Banco de Dados
}
