% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_obter_dados_qsa.R
\name{tratar_arquivo_txt}
\alias{tratar_arquivo_txt}
\title{Função trata e armazena os dados do CNPJ no Banco de Dados}
\usage{
tratar_arquivo_txt(path_arquivo_txt, localizar_cnpj, n_lines, armazenar)
}
\arguments{
\item{path_arquivo_txt}{Caminho (path) do arquivo com a base de dados do CNPJ no formato '.txt'}

\item{localizar_cnpj}{Vetor com o número dos CNPJ que se deseja filtrar e obter os dados.
O valor padrão é "NAO", o que força ao tratamento de todas as linha da base de dados}

\item{n_lines}{Número de linhas que podem ser iteradas por vez: 10000, 100000 ou 1000000}

\item{armazenar}{Indica a forma de armazenamento dos dados: 'csv' ou 'sqlite' (OBS1: O delimitador do CSV é o simbolo: "#'),
(OBS2: Preferencialmente, defina a pasta de trabalho da sessão 'Working Directory' na mesma em que está localizado o arquivo
'.txt' da base de dados no CNPJ)}
}
\description{
Função trata e armazena os dados do CNPJ no Banco de Dados
}
